ахахахах, бля, тут время в микриках в uint32 хранилось, и динамик тупо дох через 1.5 часа работы :D

+++ src/NVSSettings.h
@@ -75,6 +75,6 @@ namespace YOBA {
 
 		private:
 			constexpr static uint32_t _writeDelayTicks = pdMS_TO_TICKS(2500);
-			uint32_t _scheduledWriteTimeUs = 0;
+			int64_t _scheduledWriteTimeUs = 0;
 	};
 }
